CODE/requirements.txt
Flask==1.0.2
opencv-python==4.0.0.21
pymongo==3.7.2
keras==2.1.2
tensorflow-gpu==1.10.0
numpy==1.14.1
face_recognition==1.2.3
h5py
